Move barriers logic to src/barriers.js
